Some plots are now disabled if the needed data is missing.
Running cv or prediction will activate disabled plots.

Added a button to change the plsda_model.nr_lv and refresh plots
accordingly.

Fixed splits minimum to 2.

Fixed some widget dimensions in the right lane.

Fixed model info about RMSCEC and R^2

Still to fix RMSECV and R^2 CV TODO!

Popup about scores and loadings (and related plots) with latent
variables out of bounds [1:1] are skipped and the form is shown again.

If the test set is not compatible with the current model a popup will
notify you.

Each plot function now raises a meaningful error message (which will be
displayed in a popup) and explain the real reason with a debug message
if STATS or TEST_SET is None.